Renamed `engine.rs` to `event_loop.rs`
